+++ lib/src/utils/api_client.dart
@@ -23,11 +23,14 @@ class APIClient {
     this.mockClientHandler,
   }){
     if (mock){
+      // Mock
       client = MockClient(mockClientHandler!);
     }
     else {
+      // 真实
       client = http.Client();
     }
+    // 处理503重试
     // TODO: 通过调节RetryClient参数调整冷启动处理策略。
     client = RetryClient(client);
   }
@@ -35,6 +38,13 @@ class APIClient {
   /// 请求服务端API
   ///
   /// 遵循Python `requests`库的API风格设计。
+  ///   - 默认`Content-Type`从`text/plain`修改为`application/json`，
+  ///     以适应前后端API常见的传参习惯。
+  ///
+  /// TODO:
+  ///   - 增加特性开关`stream`参数，控制返回格式为`Response`或`StreamedResponse`，
+  ///     以方便`FutureBuilder`或者`StreamBuilder`使用。
+  ///   - 响应码为201、204等正常情况、30X等重定向情况待验证，暂不确定是否会有异常。
   Future<dynamic> requestAPI({
     required String httpMethod,
     required String apiRoot,
@@ -47,31 +57,29 @@ class APIClient {
         httpMethod,
         Uri.https(apiRoot, apiPath, queryParameters)
     );
-    // 修改默认Content-Type从text/plain为application/json
     request.headers['Content-Type'] = 'application/json';
-    // 转格式
-    request.body = json.encode(data);
+    request.body = json.encode(data);  // 转格式为String
     // 发送请求和接收响应
     http.StreamedResponse streamedResponse = await client.send(request);
-    // TODO: 增加`stream` feature flag，控制返回格式。
     http.Response response = await http.Response.fromStream(streamedResponse);
     // 处理响应
-    switch (response.statusCode){
-      case 200:
-      case 201:
-        return parseResponseData(response);
-      case 204:
-        return null;
-      case 503:
-        throw Error();
-      default:
-        throw Error();
+    if (response.statusCode<400){
+      // 正常
+      return parseResponseData(response);
+    }
+    else {
+      // 异常
+      handleHttpException(response);
     }
   }
 
-  /// 处理返回值
+  /// （暂时为非公开API）处理返回值
   ///
-  /// TODO：增加除了json格式之外的处理。
+  /// `Content-Type`为`application/json`时返回JSON解析以后的数据；
+  /// 其他情况暂时返回原始响应报文。
+  ///
+  /// TODO：
+  ///   - 允许用户继承并覆盖处理逻辑。
   dynamic parseResponseData(http.Response response){
     if (response.headers['Content-Type'] == 'application/json'){
       // https://stackoverflow.com/questions/55865173/how-to-decode-json-string-as-utf-8
@@ -81,4 +89,28 @@ class APIClient {
       return response.body;
     }
   }
+  
+  /// （暂时为非公开API）处理异常情况
+  ///
+  /// TODO：
+  ///   - 规划异常响应规范并对应实现Dart/Flutter异常处理
+  ///   - 允许用户继承并覆盖处理逻辑。
+  dynamic handleHttpException(http.Response response){
+    switch (response.statusCode) {
+      case 503:
+        // 已知问题：后台冷启动时返回
+        throw ServiceUnavailableException('服务端冷启动中', response.request?.url);
+      default:
+        throw http.ClientException('HTTP请求异常', response.request?.url);
+    }
+  }
+}
+
+
+/// 503 Service Unavailable
+class ServiceUnavailableException extends http.ClientException {
+  ServiceUnavailableException(
+      String message,
+      [Uri? uri]
+  ) : super(message, uri);
 }